Noun module upgrades: member functions and hashtable storage
Member functions (methods) are implemented in noun objects without
the need to pass a void *self argument to the method. The most
recently accessed noun is cached in a *focused global variable that
such methods can access. These member functions can then call the
static methods of the *focused object, passing *focused as the
*self argument (which the static methods still require).

Along with this, the data structure for storing and accessing noun
objects is changed to a hashtable. Previous versions used a
red-black balanced binary search tree, which was educational and
a little exotic, but not really appropriate.

Finally, Linux kernel commit message style is adopted, following
a rant by Linus Torvalds that has convinced this maintainer of
their value.

Reported-by: Jason Linehan