Noun spawning, name assignment
Adds a spawn function so that nouns can be generated
and placed at an arbitrary map location. For now, these
nouns have their name assigned by a simulated dice roll.

Adds a function that draws a house, to see if there is
any point going in that direction.

Noun methods that were useful for testing have outstayed
their welcome, so they are removed.