Fixes issue #4 (gmon.out removed from history), adds files
	AUTHORS and NEWS added

Removing gmon.out over the repository's history means
changing references to different objects in the repo,
which means this repo is not the same as the one before
this commit. There is no way to get back to the old one,
though -- even if the commits before this one are reverted
to. They're still the past, but a different past, one in
which gmon.out does not exist. It's as though I travelled
back in time, killed Hitler, and returned to the present day.
Things would be different, and looking back in time, they
would still be different from the original timeline in which
I built my time machine.